# Optimize `make_match_count_matrix` 