--- a/src/spikeinterface/comparison/comparisontools.py
+++ b/src/spikeinterface/comparison/comparisontools.py
@@ -240,10 +240,10 @@ def make_match_count_matrix(sorting1, sorting2, delta_frames, n_jobs=None):
 
     num_segments_sorting1 = sorting1.get_num_segments()
     num_segments_sorting2 = sorting2.get_num_segments()
-    max_segment_to_compare = max(num_segments_sorting1, num_segments_sorting2)
+    assert num_segments_sorting1 == num_segments_sorting2, "make_match_count_matrix : sorting1 and sorting must have the same segment number"
 
     # Segments should be matched one by one
-    for segment_index in range(max_segment_to_compare):
+    for segment_index in range(num_segments_sorting1):
         spike_vector1 = spike_vector1_segments[segment_index]
         spike_vector2 = spike_vector2_segments[segment_index]