--- a/src/spikeinterface/comparison/comparisontools.py
+++ b/src/spikeinterface/comparison/comparisontools.py
@@ -242,7 +242,7 @@ def make_match_count_matrix(sorting1, sorting2, delta_frames, n_jobs=None):
     num_segments_sorting2 = sorting2.get_num_segments()
     assert (
         num_segments_sorting1 == num_segments_sorting2
-    ), "make_match_count_matrix : sorting1 and sorting must have the same segment number"
+    ), "make_match_count_matrix : sorting1 and sorting2 must have the same segment number"
 
     # Segments should be matched one by one
     for segment_index in range(num_segments_sorting1):