… src/spikeinterface/comparison/tests/test_comparisontools.py
… src/spikeinterface/core/generate.py
… src/spikeinterface/qualitymetrics/tests/test_quality_metric_calculator.py