OK guys. This was really tedious to test and think about (I did the other Hilbert Space approach out of pure spite) but here it is. 



70 times faster in the following benchmark. It would be interesting to know if you think that some other parameters or regime is more interesting:

(`make_match_count_matrix_si` is the current method in main)
```python
durations=[18000.0], num_units=100
function.__name__='make_match_count_matrix_si'
--------------------------------------------------
Mean time over 5 iterations: 44.444810 seconds
Std over 5 iterations: 1.043592 seconds
Std without compliation over: 1.166766 seconds
Mean without compliation 44.443244 seconds
--------------------------------------------------
function.__name__='make_match_count_matrix'
--------------------------------------------------
Mean time over 5 iterations: 0.714384 seconds
Std over 5 iterations: 0.189878 seconds
Std without compliation over: 0.009023 seconds
Mean without compliation 0.619530 seconds
--------------------------------------------------
Speedup: 71.7369745920062
Speedup including compilation: 62.21417947326248
```

Here the code:
```python
num_units  = 100
#num_units = 3
durations = [5 * 60 * 60.0]
# durations = [60 * 60.0]
sampling_frequency = 30_000.0
seed = 20
sorting = generate_sorting(num_units, durations=durations, sampling_frequency=sampling_frequency, seed=seed)


samples_per_second = sampling_frequency
samples_per_millisecond = samples_per_second / 1000
delta_milliseconds = 3.334
delta_frames = int(delta_milliseconds * samples_per_millisecond)

sorting1  = sorting
sorting2  = sorting

number_it = 5

    functions = [make_match_count_matrix_si,
                 make_match_count_matrix, 
                # make_match_count_matrix_theoretical,
                 ]

    results = {}

    # Iterate through the functions and collect timing information
    for function in functions:
        print(f"{function.__name__=}")
        
        # Run the function multiple times to gather timing data
        times = [timeit.timeit(lambda: function(sorting1, sorting2, delta_frames=delta_frames), number=1) for _ in range(number_it)]
        
        # Calculate mean and std of the times
        mean_time = np.mean(times)
        std_time = np.std(times)
        mean_without_compilation = np.mean(times[1:])
        std_without_compilation = np.std(times[1:])
        
        
        print("-" * 50)
        print(f"Mean time over {number_it} iterations: {mean_time:2.6f} seconds")
        print(f"Std over {number_it} iterations: {std_time:2.6f} seconds")
        print(f"Std without compliation over: {std_without_compilation:2.6f} seconds")
        print(f"Mean without compliation {mean_without_compilation:2.6f} seconds")

        # print(f"Timing results: {times}")
        
        results[function.__name__] = {"iterations": number_it, "times": times, "mean": mean_time, "std": std_time, "mean_without_compilaton": mean_without_compilation}

        print("-" * 50)
        
    # Divide the mean without compilation of the first function by the mean without compilation of the second function
    speedup = results[functions[0].__name__]['mean_without_compilaton'] / results[functions[1].__name__]['mean_without_compilaton']
    speed_up_with_compliation = results[functions[0].__name__]['mean'] / results[functions[1].__name__]['mean']
    print(f"Speedup: {speedup}")
    print(f"Speedup with numba compilation: {speed_up_with_compliation}")
```