@alejoe91 2023-10-20
@h-mayorquin I think we need here to test `numba` import: 
https://github.com/SpikeInterface/spikeinterface/blob/main/src/spikeinterface/postprocessing/correlograms.py#L7-L12

and define the functions only `if HAVE_NUMBA`:
https://github.com/SpikeInterface/spikeinterface/blob/main/src/spikeinterface/postprocessing/correlograms.py#L323-L327

Right?