@h-mayorquin 2023-10-20
@alejoe91 
You are totally right, Alessio!

I wanted to avoid the bare try-except (as that still imports numba even when you are not using it) but for numba the type of solution that preserves the performance is a monster. Check the last commit.

I wonder if we should eat the performance importing cost and just have the simple try except.