@zm711 2023-10-20
I actually had a semi-related question to the point of this PR, so I'm hoping to just index here so if I forget I can come back. I understand the point of keeping minimal dependencies, but there is so much numba in this code why isn't it a dependency?

> There is always a quality metrics test failing somewhere, any idea about this one @alejoe91?

> It seems unrelated to the change here but I have been wrong about this before...

I've seen that test fail on my PRs before too. I think it is a non-deterministic test that fails on some runs and not others, it probably needs a patch at some point.