@samuelgarcia 2023-10-23
About dependency this is OK to more heavy dependency in sub modules pecific.
comparison having numba is OK.
what I do not like is core having numba and scipy.

So if comparison need numba we can use global import for numba I think like do evrywhere else.