@alejoe91 2023-10-24
Thanks Ramon! This looks great to me! @samuelgarcia do you have any more comments?