@samuelgarcia 2023-10-24
Let me read this again.