@h-mayorquin 2023-10-24
Just had a discussion with Sam. Some points.
1) I was not aware that sorting objects can reset their time. For that and more, spikes from different sorters should not be compared as I am currently doing. I will change this.
2) I will change to `spike_vector`, that will be improved later and I think it will make the code easier to read.
3) We are both weary of the current solution for dynamically importing numba. We should think on a general solution on how to do this.