@h-mayorquin 2023-10-31
great. lets work on the numba thing.